--- a/Gemfile
+++ b/Gemfile
@@ -5,8 +5,11 @@ source 'https://rubygems.org'
 # Specify your gem's dependencies in odata_duty.gemspec
 gemspec
 
+gem 'benchmark-ips', '>= 2.14'
 gem 'byebug'
 gem 'nokogiri'
 gem 'rake', '>= 13.0'
 gem 'rspec', '>= 3.0'
 gem 'rubocop', '>= 1.7'
+gem 'vernier', '~> 1.0'
+gem 'profile-viewer', '>= 0.0.2'