--- a/lib/odata_duty/edms.rb
+++ b/lib/odata_duty/edms.rb
@@ -50,6 +50,12 @@ module OdataDuty
       OAS_TYPE
     end
 
+    def self.to_value(object, _context)
+      object&.to_str
+    rescue StandardError => e
+      raise InvalidValue, e.message
+    end
+
     def __to_value
       object&.to_str
     rescue StandardError => e
@@ -68,6 +74,15 @@ module OdataDuty
       return { 'type' => 'array', 'items' => OAS_TYPE } if is_collection
 
       OAS_TYPE
+    end
+
+    def self.to_value(object, _context)
+      return object if object.nil?
+      return object.to_date&.iso8601 if object.respond_to?(:to_date)
+
+      Date.parse(object)&.iso8601
+    rescue StandardError => e
+      raise InvalidValue, e.message
     end
 
     def __to_value
@@ -91,6 +106,15 @@ module OdataDuty
       return { 'type' => 'array', 'items' => OAS_TYPE } if is_collection
 
       OAS_TYPE
+    end
+
+    def self.to_value(object, _context)
+      return object if object.nil?
+      return object.to_datetime&.iso8601 if object.respond_to?(:to_datetime)
+
+      DateTime.parse(object)&.iso8601
+    rescue StandardError => e
+      raise InvalidValue, e.message
     end
 
     def __to_value
@@ -117,6 +141,16 @@ module OdataDuty
     end
 
     VALID_VALUES = [true, false, nil].freeze
+
+    def self.to_value(object, _context)
+      return true if object == 'true'
+      return false if object == 'false'
+      return object if VALID_VALUES.include?(object)
+
+      object.to_boolean
+    rescue NoMethodError
+      raise InvalidValue, "#{object} not boolean value"
+    end
 
     def __to_value
       return true if object == 'true'