Before

```bash
ruby 3.3.6 (2024-11-05 revision 75015d4c1f) [x86_64-linux]
Warming up --------------------------------------
              simple    36.000 i/100ms
          odata-duty    14.000 i/100ms
Calculating -------------------------------------
              simple    332.284 (± 9.3%) i/s    (3.01 ms/i) -      1.656k in   5.032172s
          odata-duty    138.532 (±10.8%) i/s    (7.22 ms/i) -    700.000 in   5.124592s

Comparison:
              simple:      332.3 i/s
          odata-duty:      138.5 i/s - 2.40x  slower
```

After

```bash
ruby 3.3.6 (2024-11-05 revision 75015d4c1f) [x86_64-linux]
Warming up --------------------------------------
              simple    39.000 i/100ms
          odata-duty    40.000 i/100ms
Calculating -------------------------------------
              simple    411.432 (±12.2%) i/s    (2.43 ms/i) -      2.028k in   5.025922s
          odata-duty    424.624 (± 6.6%) i/s    (2.36 ms/i) -      2.120k in   5.016987s

Comparison:
          odata-duty:      424.6 i/s
              simple:      411.4 i/s - same-ish: difference falls within error
```